RFC: MapCardModule abstraction
